videos/serializers/videos/lite_video_serializer.py
from rest_framework import serializers

from videos.models import Video


class LiteVideoSerializer(serializers.ModelSerializer):

    class Meta:
        model = Video
        fields = ['id', 'title', 'slug']
