requirements.txt
Django==4.0.2
psycopg2
django-environ
djangorestframework